+++ mussel/models/phikon.py
@@ -74,7 +74,7 @@ class PhikonModel(TorchModel):
             ):
                 x = x.to(self.device, non_blocking=True)
                 out = self.obj(pixel_values=x)
-                return out.last_hidden_state[:, 0].cpu()
+                return out.last_hidden_state[:, 0].float().cpu()
 
         return model_fun
